18_lesson/requirements.txt
fastapi
fastapi-users[sqlalchemy]
uvicorn[standard]
asyncpg
Pillow
shutils